+++ src/Symfony/Bundle/FrameworkBundle/Command/WorkflowDumpCommand.php
@@ -14,6 +14,7 @@ namespace Symfony\Bundle\FrameworkBundle\Command;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Workflow\Dumper\GraphvizDumper;
 use Symfony\Component\Workflow\Dumper\StateMachineGraphvizDumper;
@@ -37,6 +38,7 @@ class WorkflowDumpCommand extends Command
             ->setDefinition(array(
                 new InputArgument('name', InputArgument::REQUIRED, 'A workflow name'),
                 new InputArgument('marking', InputArgument::IS_ARRAY, 'A marking (a list of places)'),
+                new InputOption('label', 'l', InputArgument::OPTIONAL, 'Labels a graph'),
             ))
             ->setDescription('Dump a workflow')
             ->setHelp(<<<'EOF'
@@ -73,6 +75,11 @@ EOF
             $marking->mark($place);
         }
 
-        $output->writeln($dumper->dump($workflow->getDefinition(), $marking));
+        $options = array();
+        $label = $input->getOption('label');
+        if (null !== $label && '' !== trim($label)) {
+            $options = array('graph' => array('label' => $label));
+        }
+        $output->writeln($dumper->dump($workflow->getDefinition(), $marking, $options));
     }
 }